chore: rename post components
