+++ src/components/Post/PostContentThumbnail.tsx
